refactor : FareCalulator 네이밍 및 패키지 구조 변경
